+++ tests/deck/test_deck.py
@@ -3,7 +3,7 @@
 from unittest import TestCase
 
 from hundredandten.deck import Deck
-from hundredandten.hundred_and_ten_error import HundredAndTenError
+from hundredandten.errors import HundredAndTenError
 
 
 class TestDeck(TestCase):
