+++ tests/game/test_bidding.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 
 from hundredandten.actions import Bid
 from hundredandten.constants import BidAmount, RoundRole, RoundStatus
-from hundredandten.hundred_and_ten_error import HundredAndTenError
+from hundredandten.errors import HundredAndTenError
 from hundredandten.player import remove_player_role
 from tests import arrange
