+++ tests/game/test_discard.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 
 from hundredandten.actions import Discard
 from hundredandten.constants import HAND_SIZE, RoundStatus
-from hundredandten.hundred_and_ten_error import HundredAndTenError
+from hundredandten.errors import HundredAndTenError
 from tests import arrange
