+++ tests/game/test_trump_selection.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 
 from hundredandten.actions import SelectTrump
 from hundredandten.constants import RoundStatus, SelectableSuit
-from hundredandten.hundred_and_ten_error import HundredAndTenError
+from hundredandten.errors import HundredAndTenError
 from tests import arrange
