+++ tests/suggestion/test_automated_play.py
@@ -5,7 +5,7 @@ from unittest import TestCase
 from hundredandten.actions import Action, Bid
 from hundredandten.constants import BidAmount, GameStatus
 from hundredandten.game import Game
-from hundredandten.hundred_and_ten_error import HundredAndTenError
+from hundredandten.errors import HundredAndTenError
 from hundredandten.player import NaiveAutomatedPlayer
 from tests import arrange
